requirements.txt
asgiref==3.9.1
Django==5.2.6
django-simple-history==3.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
